shams-academy/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.31.0
django-environ==0.8.1
gunicorn==21.2.0
stripe==5.0.0
python-dotenv==1.0.0